Fix a minor issue with max EP per turn
After doing an attack or action roll, the remaining
EP calculation should be correct, even if above 1x base.

Steps to reproduce:
1. Set EP to be base+2 or more. For example, if base is 10, then set to 12 or more.
2. Set boost to 1
3. Click dice roll attack

Observe that the EP becomes base
Expect that it calculates to original EP minus 1.

This appears to be simply a bug in the math that has been around for a
while.