Fix a few issues related to creating a new character (#85)
Fix If a character has no Defend skill, their Block ability is coming up as Nan. #78

When creating a new character, the dodge should be calculated correctly.
Extract a utility to ensure numbers are valid.
Add a callback to ensure that after `defend` is updated, then so is `defense`
See:
https://wiki.roll20.net/Sheet_Worker_Scripts#setAttrs.28values.2Coptions.2Ccallback.29

When `setAttrs` is called, it doesn't trigger `change` listeners. (Probably to protect from circular or infinite loops).
`setAttrs` does have a `callback` function, and I use that to fix the issue.