Update apiScripts.js
I've changed the rolling mechanism, yet again. The new one is simpler and faster (fewer rerolls, and no need to flip dice from positive to negative). It also gives an almost perfect logistic distribution. (The old one had a few too many 0s and not quite enough extreme values.) It does require a second kind of die, but that die is just an ordinary die with the 6 changed to be something else. So it is easy to simulate with an ordinary die.

I also updated the text formatting to work OK in both dark mode and light mode.