Rework top section
Summary
===
The reaction penalty is redundant with armor bonus.

The armor bonus can be negative or positive, and I often use it to manage boosts from buffs, or for negatives from stone skin or from temporary fumbles.

Removing that, opens up space for rework in the top bar.

EPN, CP, and FP don't need to clutter up the global area. Instead, the editable version should live on `Basic` and the read-only values can show up on spells and skills

Changes
===
- Remove reaction penalty
- Move EPN, CP, FP, Eff to tabs
- Move IQ, DX, BR to be a stack
- Adjust dice buttons.