Recalculate the divine spell info when the order changes.
I noticed a bug when I added the deity's name to a divine caster' spell list, made it "D", and the moved it to the top of the divine spells.
When a spell was copied to the roll section, it wasn't finding the prayer ability for the deity. It was because the reordering wasn't triggering the necessary computation.